+++ src/commonMain/kotlin/com/insanusmokrassar/sdi/HelperTools.kt
@@ -3,6 +3,7 @@ package com.insanusmokrassar.sdi
 import kotlinx.serialization.ImplicitReflectionSerializer
 import kotlinx.serialization.json.Json
 import kotlinx.serialization.modules.SerializersModuleBuilder
+import kotlin.reflect.KClass
 
 internal val nonStrictJson = Json {
     isLenient = true
@@ -11,16 +12,22 @@ internal val nonStrictJson = Json {
 }
 
 @ImplicitReflectionSerializer
-private val ModuleDeserializerStrategyWithNullOptional = ModuleDeserializerStrategy()
-
-@ImplicitReflectionSerializer
 fun Json.loadModule(
     json: String,
+    vararg additionalClassesToInclude: KClass<*>,
     moduleBuilder: (SerializersModuleBuilder.() -> Unit)? = null
-): Module = parse(if (moduleBuilder != null) ModuleDeserializerStrategy(moduleBuilder) else ModuleDeserializerStrategyWithNullOptional, json)
+): Module = parse(
+    if (moduleBuilder != null) {
+        ModuleDeserializerStrategy(moduleBuilder, *additionalClassesToInclude)
+    } else {
+        ModuleDeserializerStrategy(additionalClassesToInclude = *additionalClassesToInclude)
+    },
+    json
+)
 
 @ImplicitReflectionSerializer
 fun loadModule(
     json: String,
+    vararg additionalClassesToInclude: KClass<*>,
     moduleBuilder: (SerializersModuleBuilder.() -> Unit)? = null
-): Module = nonStrictJson.loadModule(json, moduleBuilder)
+): Module = nonStrictJson.loadModule(json, *additionalClassesToInclude, moduleBuilder = moduleBuilder)